requirements.txt
pyspark==3.5.0
pandas==2.2.2
joblib==1.4.2
findspark==2.0.1
streamlit
pandas
matplotlib
psycopg2-binary
kafka-python
streamlit-extras
folium
pydeck